// smoke-effect/src/app/shared/models/products.model.ts
export interface IProducts {
    id: number;
    name: string;
    price: number;
    section: string;
    imageUrl?: string;
    brandName: string;
    type: string;
     stock: number;
    origin: string;
    evaluation: number;
    condition: number;
    strength: number;
    want: number;
    sold: number;
}